+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @idinghuo/hooks
 
+## 2.3.0
+
+### Minor Changes
+
+- 6
+
+## 2.2.0
+
+### Minor Changes
+
+- 5
+
+## 2.1.0
+
+### Minor Changes
+
+- 3
+
+## 2.0.0
+
+### Major Changes
+
+- 大版本
+
+### Minor Changes
+
+- 发版
+- 2
+
+## 1.2.0
+
+### Minor Changes
+
+- 再次尝试推包
+
 ## 1.1.0
 
 ### Minor Changes